requirements.txt
flask==2.3.3
pymongo[srv]==4.6.1
python-dotenv==1.0.1
dnspython==2.4.2
gunicorn==21.2.0
